+++ lib/src/localizations/mixins/ordinal_numbers.dart
@@ -4,6 +4,9 @@ import 'package:moment_dart/moment_dart.dart';
 mixin Ordinal on MomentLocalization {
   String ordinalNumber(int n);
 
+  /// Overrides:
+  ///
+  /// `Mo`, `Qo`, `Do`, `DDDo`, `d_o`, and `wo`
   Map<FormatterToken, FormatterTokenFn> get formattersWithOrdinal => {
         FormatterToken.Mo: (DateTime dateTime) => ordinalNumber(dateTime.month),
         FormatterToken.Qo: (DateTime dateTime) =>